Struct entitypool::EntityPool

pub struct EntityPool {
    // some fields omitted
}

Methods

impl EntityPool

fn new() -> EntityPool

Creates a new, empty EntityPool.

The EntityPool will not allocate until entities are created.

fn with_capacity(create_capacity: usize, return_capacity: usize) -> EntityPool

Creates a new, empty EntityPool with the specified capacities.

The EntityPool will able to create create_capacity and return return_capacity entities without reallocating. If either capacity is 0, their respective storage vectors will not allocate.

fn create_entity(&mut self) -> (usize, Entity)

Creates a unique entity.

Returns the Entity and it's current index. The index is only guaranteed to remain valid until the next call to return_entity.

fn return_entity(&mut self, entity: Entity)

Release ownership of the entity, allowing for it to be recycled. A recycled entity will have it's internal generation incremented, yielding a new, unique entity.

Entities are stored in contiguous memory. When an entity is returned, the last entity is swaped into the returned entity's slot; thus indexes retrieved prior to returning an entity are potentially invalidated.

Panics

Returning an entity more than once, or an entity created from another pool, results in undefined behavior.

Examples

use entitypool::EntityPool;

let mut pool = EntityPool::new();
let (_, e1) = pool.create_entity();
pool.return_entity(e1);
let (_, e2) = pool.create_entity();
assert!(e1 != e2);

use entitypool::EntityPool;

let mut pool = EntityPool::new();
let (i1, e1) = pool.create_entity();
let (i2, e2) = pool.create_entity();
let (i3, e3) = pool.create_entity();
assert_eq!(0, i1);
assert_eq!(1, i2);
assert_eq!(2, i3);
assert_eq!(0, pool.index_of(e1));
assert_eq!(1, pool.index_of(e2));
assert_eq!(2, pool.index_of(e3));
pool.return_entity(e2);
assert_eq!(0, pool.index_of(e1));
assert_eq!(1, pool.index_of(e3));

fn index_of(&self, entity: Entity) -> usize

Returns the current index of the given entity, which is only guaranteed to remain valid until the next call to return_entity.

Panics

Querying the status of an entity from another pool results in undefined behavior.

fn entity_at(&self, index: usize) -> Entity

Returns the current entity at the given index.

Panics

Panics if the index is greater or equal to the number of live entities in this pool.

fn is_alive(&self, entity: Entity) -> bool

Returns true if this entity has not been returned.

Panics

Querying the status of an entity from another pool results in undefined behavior.

fn iter(&self) -> Iter<Entity>

Returns an iterator to the live entities. The Enumerate of the returned iterator will yield each entity and its current index.

Examples

use entitypool::EntityPool;

let mut pool = EntityPool::new();
pool.create_entity();
pool.create_entity();
pool.create_entity();
for (index, entity) in pool.iter().enumerate() {
    assert_eq!(index, pool.index_of(*entity));
    assert_eq!(*entity, pool.entity_at(index));
}

fn reset(&mut self)

Resets the EnitityPool to its initial state, without releasing allocated capacity.

All entities created prior to resetting are no longer considered members of this pool.

fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more entities to be created without reallocation. The pool may reserve more space to avoid frequesnt reallocations.

fn reserve_returned(&mut self, additional: usize)

Reserves capacity for at least additional more entities to be returned without reallocation. The pool may reserve more space to avoid frequesnt reallocations.

fn shrink_to_fit(&mut self)

Shrinks the capacity of this pool as much as possible.

fn len(&self) -> usize

Returns the number of live entities in this pool.

fn len_returned(&self) -> usize

Returns the number of returned entities in this pool that are ready to be recycled.

fn capacity(&self) -> usize

Returns the number of entities that this pool can create without reallocation.

fn capacity_returned(&self) -> usize

Returns the number of entities that can be returned without reallocation.